Move hook creation to file in src/lib/controller
