+++ src/lib/controller/createHooks.ts
@@ -1,10 +1,10 @@
+import { ControllerHooks } from ".";
 import { resolveIgnoreNamespaces } from "../assets/webhooks";
-import { ControllerHooks } from "../controller";
+import { Capability } from "../core/capability";
+import { isWatchMode, isDevMode } from "../core/envChecks";
 import { setupWatch } from "../processors/watch-processor";
 import Log from "../telemetry/logger";
 import { PeprModuleOptions } from "../types";
-import { Capability } from "./capability";
-import { isWatchMode, isDevMode } from "./envChecks";
 
 /**
  * Creates controller hooks with proper handling of watch setup
